+++ changelog.md
@@ -2,6 +2,11 @@
 
 ---
 
+## vNext
+
+### Fixed
+- Fixed issue where the static bucket url in the Cloudformation Output was wrong for newer AWS regions; thanks @thedersen
+
 ## [6.0.2] 2020-12-04
 
 ### Changed